+++ src/AbstractEntity.php
@@ -10,6 +10,12 @@
 
 namespace WNowicki\Generic;
 
+use WNowicki\Generic\Contracts\Arrayable;
+use WNowicki\Generic\Contracts\Entity;
+use WNowicki\Generic\Contracts\Jsonable;
+use WNowicki\Generic\Contracts\Makeable;
+use WNowicki\Generic\Exceptions\InvalidArgumentException;
+
 /**
  * Abstract Entity
  *
@@ -18,8 +24,14 @@ namespace WNowicki\Generic;
  *
  * @package WNowicki\Generic
  */
-abstract class AbstractEntity implements EntityInterface
+abstract class AbstractEntity implements Entity, Makeable, Jsonable
 {
+    const TYPE_ARRAY = 1;
+    const TYPE_BOOL = 2;
+    const TYPE_INT = 4;
+    const TYPE_STRING = 8;
+    const TYPE_FLOAT = 16;
+
     private $data = [];
 
     protected $properties = [];
@@ -39,7 +51,7 @@ abstract class AbstractEntity implements EntityInterface
 
             if ($entity->isPropertyAllowed($k)) {
 
-                $entity->set($k, [$v]);
+                $entity->{'set' . $entity->snakeToCamel($k)}($v);
             }
         }
 
@@ -55,7 +67,7 @@ abstract class AbstractEntity implements EntityInterface
     public function __call($name, $arguments)
     {
         $action = substr($name, 0, 3);
-        $property = $this->nameConverter(substr($name, 3));
+        $property = $this->camelToSnake(substr($name, 3));
 
         if ($this->isPropertyAllowed($property)) {
 
@@ -65,52 +77,37 @@ abstract class AbstractEntity implements EntityInterface
                 case 'get':
                     return $this->get($property);
             }
+// @codeCoverageIgnoreStart
         }
+// @codeCoverageIgnoreEnd
 
         trigger_error('Call to undefined method '.__CLASS__.'::'.$name.'()', E_USER_ERROR);
+// @codeCoverageIgnoreStart
     }
+// @codeCoverageIgnoreEnd
 
     /**
      * To Array
      *
-     * This implementation handles automatically Scalars, Arrays and EntityInterface. Everything else is ignored.
-     * Return flatten (arrays of scalars (+ null)???) representation of Entity
-     *
      * @author WN
+     * @param bool $recursively If set to `true` then toArray(true) will be called on each `Arrayable` property
      * @return array
      */
-    public function toArray()
+    public function toArray($recursively = false)
     {
         $rtn = [];
 
         foreach ($this->data as $k => $v) {
-            $this->flattenProperty($k, $v, $rtn);
+
+            if ($recursively && $v instanceof Arrayable) {
+                $rtn[$k] = $v->toArray(true);
+                continue;
+            }
+
+            $rtn[$k] = $v;
         }
 
         return $rtn;
-    }
-
-    /**
-     * @author WN
-     * @param string $k
-     * @param mixed $v
-     * @param array $rtn
-     * @return null
-     */
-    private function flattenProperty($k, $v, array &$rtn)
-    {
-        if (is_scalar($v) || is_array($v)) {
-
-            $rtn[$k] = $v;
-            return null;
-        }
-
-        if ($v instanceof EntityInterface) {
-
-            $rtn[$k] = $v->toArray();
-        }
-
-        return null;
     }
 
     /**
@@ -120,7 +117,11 @@ abstract class AbstractEntity implements EntityInterface
      */
     private function isPropertyAllowed($property)
     {
-        return (count($this->properties) == 0 || in_array($property, $this->properties));
+        return (
+            count($this->properties) == 0 ||
+            in_array($property, $this->properties) ||
+            array_key_exists($property, $this->properties)
+        );
     }
 
     /**
@@ -129,7 +130,18 @@ abstract class AbstractEntity implements EntityInterface
      */
     public function __toString()
     {
-        return json_encode($this->toArray());
+        return $this->toJson();
+    }
+
+    /**
+     * JSON representation of an object
+     *
+     * @param  int $options
+     * @return string
+     */
+    public function toJson($options = 0)
+    {
+        return json_encode($this->toArray(true), $options);
     }
 
     /**
@@ -138,14 +150,16 @@ abstract class AbstractEntity implements EntityInterface
      * @param array $arguments
      * @return $this
      */
-    protected function set($property, $arguments)
+    private function set($property, $arguments)
     {
         if (count($arguments) == 0) {
 
             trigger_error('Missing argument on method ' . __CLASS__ . '::set_' . $property . '() call', E_USER_ERROR);
+// @codeCoverageIgnoreStart
         }
+// @codeCoverageIgnoreEnd
 
-        $this->data[$property] = $arguments[0];
+        $this->data[$property] = $this->processInputValue($arguments[0], $property);
 
         return $this;
     }
@@ -167,12 +181,161 @@ abstract class AbstractEntity implements EntityInterface
 
     /**
      * @author WN
-     * @param string $name
+     * @param mixed $value
+     * @param string|int $property
+     * @return mixed
+     * @throws Exception
+     * @throws InvalidArgumentException
+     */
+    private function processInputValue($value, $property)
+    {
+        if (array_key_exists($property, $this->properties)) {
+            $type = $this->properties[$property];
+            if ($this->isInternalType($type)) {
+
+                return $this->processInternalType($value, $type);
+            }
+
+            return $this->processObjectType($value, $type);
+        }
+
+        return $value;
+    }
+
+    /**
+     * @author WN
+     * @param $type
+     * @return bool
+     */
+    private function isInternalType($type)
+    {
+        return is_numeric($type) && array_key_exists($type, $this->propertyInternalTypes());
+    }
+
+    /**
+     * @author WN
+     * @param mixed $value
+     * @param int $type
+     * @return int|float|bool|string|array
+     * @throws InvalidArgumentException
+     */
+    private function processInternalType($value, $type)
+    {
+        if ($this->validateInternalType($value, $type)) {
+
+            return $type;
+        }
+
+        throw new InvalidArgumentException(
+            'Expected value to be type of [' . $this->propertyInternalTypes()[$type] .
+            '] type [' . $this->checkType($value) . '] was given'
+        );
+    }
+
+    /**
+     * @param mixed $value
+     * @param string $class
+     * @return object mixed
+     * @throws Exception
+     * @throws InvalidArgumentException
+     */
+    private function processObjectType($value, $class)
+    {
+        $this->classExists($class);
+
+        if (is_array($value) && is_subclass_of($class, 'WNowicki\Generic\Contracts\Makeable')) {
+
+            return $class::make($value);
+        }
+
+        if (is_a($value, $class)) {
+
+            return $value;
+        }
+
+        throw new InvalidArgumentException(
+            'Expected value to be object of [' . $class . '] type ' . $this->checkType($value) . '] was given'
+        );
+    }
+
+    private function classExists($class)
+    {
+        if (class_exists($class)) {
+            return true;
+        }
+        throw new Exception('Non existing class');
+    }
+
+    /**
+     * @author WN
+     * @param $value
+     * @param $type
+     * @return bool|null
+     */
+    private function validateInternalType($value, $type)
+    {
+        switch ($type) {
+            case self::TYPE_ARRAY:
+                return is_array($value);
+            case self::TYPE_INT:
+                return is_int($value);
+            case self::TYPE_STRING:
+                return is_string($value);
+            case self::TYPE_BOOL:
+                return is_bool($value);
+            case self::TYPE_FLOAT:
+                return is_float($value);
+        }
+    }
+
+    /**
+     * @author WN
+     * @return array
+     */
+    private function propertyInternalTypes()
+    {
+        return [
+            self::TYPE_ARRAY => 'array',
+            self::TYPE_INT => 'int',
+            self::TYPE_STRING => 'string',
+            self::TYPE_BOOL => 'bool',
+            self::TYPE_FLOAT => 'float',
+        ];
+    }
+
+    /**
+     * @author WN
+     * @param mixed $value
      * @return string
      */
-    private function nameConverter($name)
+    private function checkType($value)
     {
-        $name[0] = strtolower($name[0]);
-        return strtolower(preg_replace("/([A-Z])/", "_$1", $name));
+        if (is_object($value)) {
+
+            return get_class($value);
+        }
+
+        return gettype($value);
+    }
+
+    /**
+     * @author WN
+     * @param string $string
+     * @return string
+     */
+    private function camelToSnake($string)
+    {
+        $string[0] = strtolower($string[0]);
+        return strtolower(preg_replace("/([A-Z])/", "_$1", $string));
+    }
+
+    /**
+     * @author WN
+     * @param $string
+     * @return mixed
+     */
+    private function snakeToCamel($string)
+    {
+        return str_replace(' ', '', ucwords(str_replace('_', ' ', $string)));
     }
 }